implement gamewatchlist delete btn functonanlity

+++ src/components/alert/deleteAlert.jsx
@@ -1,12 +1,16 @@
 import Swal from "sweetalert2";
 
-export const deleteAlert = (id, data) => {
-  const { myReviewList, setMyReviewList } = data;
+export const deleteAlert = (id, data, condition) => {
+  const { myReviewList, setMyReviewList, watchListIDs, setWatchListIDs, user } =
+    data;
   const swalWithBootstrapButtons = Swal.mixin({
     customClass: {
       confirmButton: "btn btn-success text-white m-2",
       cancelButton: "btn btn-error text-white",
       cancelButtonColor: "",
+      background: "#4d4d4d",
+      backdrop: "#b8f0c858",
+      color: "white",
     },
     buttonsStyling: false,
   });
@@ -19,34 +23,73 @@ export const deleteAlert = (id, data) => {
       confirmButtonText: "Yes, delete it!",
       cancelButtonText: "No, cancel!",
       reverseButtons: true,
+      background: "#4d4d4d",
+      backdrop: "#b8f0c858",
+      color: "white",
     })
     .then((result) => {
       if (result.isConfirmed) {
-        console.log("delete");
+        if (condition === "formReview") {
+          fetch("http://localhost:4000/reviews", {
+            method: "DELETE",
+            headers: { "content-type": "application/json" },
+            body: JSON.stringify({ id: id }),
+          })
+            .then((res) => res.json())
+            .then((data) => {
+              if (data.deletedCount > 0) {
+                const remain = myReviewList.filter((list) => list._id !== id);
+                setMyReviewList(remain);
+                swalWithBootstrapButtons.fire({
+                  title: "Deleted!",
+                  text: "Your file has been deleted.",
+                  icon: "success",
+                  background: "#4d4d4d",
+                  backdrop: "#b8f0c858",
+                  color: "white",
+                });
+              }
+            });
+        }
+        if (condition === "formWatch") {
+          const remainReviews = watchListIDs.filter((list) => list._id !== id);
+          const remainIds = remainReviews.map((id) => id._id);
+          console.log(remainIds);
+          const updateWatchListData = {
+            email: user.email,
+            user: user.name,
+            ids: remainIds,
+            isComplete: false,
+          };
 
-        fetch("http://localhost:4000/reviews", {
-          method: "DELETE",
-          headers: { "content-type": "application/json" },
-          body: JSON.stringify({ id: id }),
-        })
-          .then((res) => res.json())
-          .then((data) => {
-            if (data.deletedCount > 0) {
-              const remain = myReviewList.filter((list) => list._id !== id);
-              setMyReviewList(remain);
-              swalWithBootstrapButtons.fire({
-                title: "Deleted!",
-                text: "Your file has been deleted.",
-                icon: "success",
-              });
-            }
-          });
+          fetch(`http://localhost:4000/updateWatchList/${user.email}`, {
+            method: "PUT",
+            headers: { "content-type": "application/json" },
+            body: JSON.stringify(updateWatchListData),
+          })
+            .then((res) => res.json())
+            .then((data) => {
+              if (data.modifiedCount > 0) {
+                setWatchListIDs(remainReviews);
+                swalWithBootstrapButtons.fire({
+                  title: "Deleted!",
+                  text: "Your file has been deleted.",
+                  icon: "success",
+                  background: "#4d4d4d",
+                  backdrop: "#b8f0c858",
+                  color: "white",
+                });
+              }
+            });
+        }
       } else if (result.dismiss === Swal.DismissReason.cancel) {
-        console.log("cancel");
         swalWithBootstrapButtons.fire({
           title: "Cancelled",
           text: "Your imaginary file is safe :)",
           icon: "error",
+          background: "#4d4d4d",
+          backdrop: "#b8f0c858",
+          color: "white",
         });
       }
     });
